+++ CHANGELOG.md
@@ -7,6 +7,9 @@
 * Refine automatically collected breadcrumbs to a commonly useful set by default
 [#321](https://github.com/bugsnag/bugsnag-android/pull/321)
 
+* Ensure that unhandled error reports are always sent immediately on launch for Android P and in situations with no connectivity.
+[#319](https://github.com/bugsnag/bugsnag-android/pull/319)
+
 ## 4.4.0 (2018-05-17)
 
 ### Features