+++ CHANGELOG.md
@@ -1,8 +1,11 @@
 # Changelog
 
-## 4.4.1 (TBD)
+## TBD
 
 ### Bug fixes
+
+* Refine automatically collected breadcrumbs to a commonly useful set by default
+[#321](https://github.com/bugsnag/bugsnag-android/pull/321)
 
 * Ensure that unhandled error reports are always sent immediately on launch for Android P and in situations with no connectivity.
 [#319](https://github.com/bugsnag/bugsnag-android/pull/319)